# core: shard SerializeExecutor queues