Motivation:
Today: each Call is passed an Executor to run its callbacks on.
This executor is wrapped in a SerializingExecutor (SE) which ensures
all callbacks happen in a serial fashion, and on the provded
executor.

This has a number of drawbacks, in that each SE has to maintain its
own queue of runnables, and is likely to schedule work on an idle
Executor thread.

Additionally, the SE API doesn't lend itself to reuse, notably in
that it doesn't handle direct executors well.

Modification:

* SE has been split into a shard called a SingleSerializingExecutor
* SE is now an interface that extends Executor.
* A new class ShardingSerializingExecutor shards access to
  SingleSerializingExecutor
* A new SerializingExecutors class is responsible for construction,
  including returning the proper type for direct executors.
* ManagedChannelImpl and ServerImpl now have a single sharded SE
  (called a SerializingExecutorFactory) which doles out SE.

Alternatives Considered:

Sharding inside the old SE has some problems.  It is specific to
an executor, so it can't share queues globally.  Effectively this
means each executor has to have its own sharded queues.

SE Was a class, but now it is an interface.  Callers of new SE()
always had to get the top level implementation, and there was no
room for returning a more specific type.  It also doesn't lend
itself to reusing itself.  Making SE an abstract class forces
an implementation, but it may be inconvenient to use.  Since
Executor is an interface SE should be one too.

Learning from the API design of ManagedChannel.builder(), the
factory for SEs is not on the SE class itself, but on a side
class.  This avoids child classes inheriting a constructor that
is not applicable.

The default sharding amount is 1 instead of num processors.
As the number of queues increases, performance gets worse.
1 is not the right number because none of our benchmarks
test when under high application load.  Clients with high load
will need to create multiple channels with this currently.

Results:
Overall performance is slightly better, with some benchmarks
being a wash.

* The TransportBenchmark suite, which measures
  blocking stub latency is mostly a wash.  This is because
  blocking stubs use their own Threadless executor, forcing a new
  SE each time.  This is unfortunate, but can be fixed in the
  future.

* The QpsClient and QpsServer benchmarks show noticeable gain,
  improving throughput and latency considerably.  I ran the
  benchmarks alternativingly a few times to make sure they
  were not a fluke.  The gain here is due to the callbacks
  ending up on the same queue.  I believe this results in less
  thread contention.

* FlowControlledMessagesPerSecond is a minor bit better for the
  new implementation.

Data:
```
BEFORE:
Benchmark                                               (direct)  (transport)    Mode      Cnt         Score     Error  Units
TransportBenchmark.unaryCall1024                            true    INPROCESS  sample  1162772      4382.065 ±  18.697  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00        true    INPROCESS  sample               1714.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50        true    INPROCESS  sample               4176.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90        true    INPROCESS  sample               5040.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95        true    INPROCESS  sample               5328.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99        true    INPROCESS  sample               9700.320            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999       true    INPROCESS  sample              18016.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999      true    INPROCESS  sample              62620.506            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00        true    INPROCESS  sample            3448832.000            ns/op
TransportBenchmark.unaryCall1024                            true        NETTY  sample   577543     69197.871 ± 281.233  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00        true        NETTY  sample              57664.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50        true        NETTY  sample              65152.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90        true        NETTY  sample              80128.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95        true        NETTY  sample              87552.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99        true        NETTY  sample             100480.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999       true        NETTY  sample             140032.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999      true        NETTY  sample             693642.854            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00        true        NETTY  sample           23920640.000            ns/op
TransportBenchmark.unaryCall1024                            true  NETTY_LOCAL  sample   748664     53346.942 ± 115.913  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00        true  NETTY_LOCAL  sample              38592.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50        true  NETTY_LOCAL  sample              50944.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90        true  NETTY_LOCAL  sample              61440.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95        true  NETTY_LOCAL  sample              66688.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99        true  NETTY_LOCAL  sample              79104.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999       true  NETTY_LOCAL  sample             110762.880            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999      true  NETTY_LOCAL  sample             405914.112            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00        true  NETTY_LOCAL  sample           11845632.000            ns/op
TransportBenchmark.unaryCall1024                            true       OKHTTP  sample   610687     65406.705 ± 268.297  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00        true       OKHTTP  sample              53056.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50        true       OKHTTP  sample              62848.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90        true       OKHTTP  sample              73856.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95        true       OKHTTP  sample              79488.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99        true       OKHTTP  sample              97152.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999       true       OKHTTP  sample             132943.872            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999      true       OKHTTP  sample             399219.098            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00        true       OKHTTP  sample           26705920.000            ns/op
TransportBenchmark.unaryCall1024                           false    INPROCESS  sample  1220966      8209.184 ±  27.438  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00       false    INPROCESS  sample               3576.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50       false    INPROCESS  sample               7704.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90       false    INPROCESS  sample              10656.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95       false    INPROCESS  sample              10992.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99       false    INPROCESS  sample              13104.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999      false    INPROCESS  sample              25952.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999     false    INPROCESS  sample              82419.622            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00       false    INPROCESS  sample            4243456.000            ns/op
TransportBenchmark.unaryCall1024                           false        NETTY  sample   475120     84078.144 ± 339.944  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00       false        NETTY  sample              63808.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50       false        NETTY  sample              79744.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90       false        NETTY  sample              95232.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95       false        NETTY  sample             102272.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99       false        NETTY  sample             118784.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999      false        NETTY  sample             166625.024            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999     false        NETTY  sample            1141735.219            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00       false        NETTY  sample           23822336.000            ns/op
TransportBenchmark.unaryCall1024                           false  NETTY_LOCAL  sample   589551     67751.608 ±  94.927  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00       false  NETTY_LOCAL  sample              45184.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50       false  NETTY_LOCAL  sample              64704.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90       false  NETTY_LOCAL  sample              76800.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95       false  NETTY_LOCAL  sample              81536.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99       false  NETTY_LOCAL  sample              96640.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999      false  NETTY_LOCAL  sample             147314.688            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999     false  NETTY_LOCAL  sample            1132635.750            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00       false  NETTY_LOCAL  sample            4440064.000            ns/op
TransportBenchmark.unaryCall1024                           false       OKHTTP  sample   489258     81651.028 ± 339.245  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00       false       OKHTTP  sample              59136.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50       false       OKHTTP  sample              78848.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90       false       OKHTTP  sample              91136.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95       false       OKHTTP  sample              96896.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99       false       OKHTTP  sample             114304.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999      false       OKHTTP  sample             156672.000            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999     false       OKHTTP  sample            1136791.757            ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00       false       OKHTTP  sample           25788416.000            ns/op

AFTER:

TransportBenchmark.unaryCall1024                            true    INPROCESS  sample  1052629      5037.509 ±  336.117  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00        true    INPROCESS  sample               1724.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50        true    INPROCESS  sample               4352.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90        true    INPROCESS  sample               5216.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95        true    INPROCESS  sample               5448.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99        true    INPROCESS  sample              11136.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999       true    INPROCESS  sample              19831.680             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999      true    INPROCESS  sample              74300.672             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00        true    INPROCESS  sample           32014336.000             ns/op
TransportBenchmark.unaryCall1024                            true        NETTY  sample   579150     68968.436 ±  300.382  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00        true        NETTY  sample              55360.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50        true        NETTY  sample              63424.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90        true        NETTY  sample              80512.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95        true        NETTY  sample              88320.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99        true        NETTY  sample              99328.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999       true        NETTY  sample             159961.344             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999      true        NETTY  sample            1749339.750             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00        true        NETTY  sample           15482880.000             ns/op
TransportBenchmark.unaryCall1024                            true  NETTY_LOCAL  sample   712680     52533.468 ±  270.518  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00        true  NETTY_LOCAL  sample              38080.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50        true  NETTY_LOCAL  sample              49856.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90        true  NETTY_LOCAL  sample              61696.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95        true  NETTY_LOCAL  sample              67712.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99        true  NETTY_LOCAL  sample              80512.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999       true  NETTY_LOCAL  sample             117928.832             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999      true  NETTY_LOCAL  sample             844489.523             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00        true  NETTY_LOCAL  sample           49217536.000             ns/op
TransportBenchmark.unaryCall1024                            true       OKHTTP  sample   615898     64844.861 ±  276.037  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00        true       OKHTTP  sample              51840.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50        true       OKHTTP  sample              61376.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90        true       OKHTTP  sample              72576.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95        true       OKHTTP  sample              77952.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99        true       OKHTTP  sample              94080.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999       true       OKHTTP  sample             140569.856             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999      true       OKHTTP  sample            2522399.539             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00        true       OKHTTP  sample           14548992.000             ns/op
TransportBenchmark.unaryCall1024                           false    INPROCESS  sample   972517     10536.111 ±  350.760  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00       false    INPROCESS  sample               4144.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50       false    INPROCESS  sample               9984.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90       false    INPROCESS  sample              10912.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95       false    INPROCESS  sample              11392.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99       false    INPROCESS  sample              14064.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999      false    INPROCESS  sample              24736.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999     false    INPROCESS  sample             105758.387             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00       false    INPROCESS  sample           36634624.000             ns/op
TransportBenchmark.unaryCall1024                           false        NETTY  sample   432355     92429.816 ± 2968.356  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00       false        NETTY  sample              64320.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50       false        NETTY  sample              80896.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90       false        NETTY  sample              96896.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95       false        NETTY  sample             103680.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99       false        NETTY  sample             120832.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999      false        NETTY  sample             228864.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.9999     false        NETTY  sample           48893630.874             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p1.00       false        NETTY  sample           65208320.000             ns/op
TransportBenchmark.unaryCall1024                           false  NETTY_LOCAL  sample   598409     66740.930 ±  440.637  ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.00       false  NETTY_LOCAL  sample              41088.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.50       false  NETTY_LOCAL  sample              63424.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.90       false  NETTY_LOCAL  sample              75392.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.95       false  NETTY_LOCAL  sample              80000.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.99       false  NETTY_LOCAL  sample              94848.000             ns/op
TransportBenchmark.unaryCall1024:unaryCall1024·p0.999      false  NETTY_LOCAL  sample             144384.000             ns/op
…