@ejona86 2017-09-05
> in that each SE has to maintain its
own queue of runnables, and is likely to schedule work on an idle
Executor thread.

Why is this a problem? That sounds good to me.

> Additionally, the SE API doesn't lend itself to reuse, notably in
that it doesn't handle direct executors well.

What is this referring to?

I just took a glance, but it seems like this uses a single SerializingExecutor for everything (I see 1 shard specified it `ManagedChannelImpl`). Yes, we can change it to use multiple shards, but we are still forced to figure out that configuration.

I see two problems with the approach:
* We don't know the number of shards to use. We don't know the concurrency level the application requires. This is why we use a cached thread pool instead of a fixed-size thread pool by default.
* This can starve callbacks, if they are placed on the same shard as a callback that blocks for a very long time.

Today it is fine to block in callbacks, but with this change it would not be. Yes, the QPS benchmark is better, but it does no work in the application threads. In a real application that sometimes blocks, I would expect tail latency to increase dramatically.