@carl-mastrangelo 2017-09-08
@ejona86 You are right, I have removed all the sharding logic, but kept the factory and interfaces.  

> What is this referring to?

In some of the code, we check if the passed in executor is a directExecutor.  This consolidates the logic, by hiding the SE construction in `SerializingExecutors`.  Also, it avoids wrapping an already serializing exec.

Notably, the `ThreadlessExecutor` in ClientCalls is now marked SE, so it doesn't get wrapped for each RPC.  This saves an SE alloc. 

PTAL